requirements.txt
Pillow>=10.0.0
python-docx>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
